Refactor header component
